+++ tests/testthat/test-tidyeval-across.R
@@ -470,7 +470,7 @@ test_that("if_all/any is wrapped in parentheses #1153", {
 
   expect_equal(
     lf |> filter(if_any(c(a, b)) & c == 3) |> remote_query(),
-    sql("SELECT `df`.*\nFROM `df`\nWHERE ((`a` OR `b`) AND `c` = 3.0)")
+    sql("SELECT *\nFROM `df`\nWHERE ((`a` OR `b`) AND `c` = 3.0)")
   )
 })
 
@@ -514,11 +514,11 @@ test_that("if_any() and if_all() expansions deal with single inputs", {
   # Single inputs
   expect_equal(
     filter(d, if_any(x, ~FALSE)) |> remote_query(),
-    sql("SELECT `df`.*\nFROM `df`\nWHERE ((FALSE))")
+    sql("SELECT *\nFROM `df`\nWHERE ((FALSE))")
   )
   expect_equal(
     filter(d, if_all(x, ~FALSE)) |> remote_query(),
-    sql("SELECT `df`.*\nFROM `df`\nWHERE ((FALSE))")
+    sql("SELECT *\nFROM `df`\nWHERE ((FALSE))")
   )
 })
 
@@ -635,12 +635,12 @@ test_that("can `arrange()` with `pick()` selection", {
 
   expect_identical(
     arrange(df, pick(x, y)) |> remote_query(),
-    sql("SELECT `df`.*\nFROM `df`\nORDER BY `x`, `y`")
+    sql("SELECT *\nFROM `df`\nORDER BY `x`, `y`")
   )
 
   expect_identical(
     arrange(df, pick(x), y) |> remote_query(),
-    sql("SELECT `df`.*\nFROM `df`\nORDER BY `x`, `y`")
+    sql("SELECT *\nFROM `df`\nORDER BY `x`, `y`")
   )
 })